Fix marker popup on budget investments
Since we're directly writing the URL in the JavaScript, the relative URL
root wasn't automatically included. Usually we'd add a `data` attribute
so we can get the URL from the JavaScript without duplicating it; we
aren't bothering to do so because we aren't going to use participatory
budgets for now.