Use a custom component to render a custom ERB file
In commit a2ec8d3cc, a component ERB file from CONSUL was temporarily
changed.

Since this was a temporary commit, we're reverting it. We're using a
custom component instead so ViewComponent finds the ERB file in the
custom folder.

We're also updating the test so it checks the arrow buttons don't show
up, as mentioned in commit 6bac19e18.