Add no longer used local_postal_code method
We're adding it again because it uses a translation that is used nowhere
else, and i18n-tasks was reporting it as unused.

We could add this translation to the `ignore_unused` ones, but since
version 1.5 of CONSUL will move this method to the "standard"
Verification::Residence model, we're keeping the method instead. We'll
remove it again when upgrading to CONSUL 1.5.