Simplify tests for the sign in action
We've already got tests covering the sign in link in the `sessions_spec`
file. In other tests, we can go directly to the sign in page, making the
tests slightly shorter and faster.

We also make it easier to customize the "sign in" link; CONSUL
installations customizing its text (or even its presence) won't have to
change so many tests.