Fix translation interface with jQuery 3
Quoting the jQuery Core 3.0 Upgrade Guide:

> Breaking change: jQuery("#") and .find("#") are invalid syntax

This was internally used by Foundation when providing an empty hash as
the link to the tab. Upgrading Foundation didn't solve the issue.

As a workaround, I'm generating a useless hash instead of an empty one.

Note we didn't have this issue when updating jQuery in Consul because
Consul changed the translation interface before updating to jQuery 3.