Assign roles when authenticating via SAML
We're using the names of the LDAP groups created by the Government of
the Canary Islands, like `gaut_admin_ecociv_consul` for administrators.

We've tried to add the `isMemberOf` attribute to the `auth.info` hash,
but when addid it by configuring `request_attributes` or
`attribute_statements`, we were only getting one role instead of all the
roles that are present in the `isMemberOf` array.

Note that, if the roles of a user change, this will only take effect
after signing out and signing in again. We think it's reasonable since
groups won't change often, and the requirement to sign in again is
common in Unix systems when adding a user to a group.