Allow selecting the time when a poll starts/ends
We were already saving it as a time, but we didn't offer an interface to
select the time due to lack of decent browser support for this field
back when this feature was added.

However, nowadays all major browsers support this field type and, at the
time of writing, at least 86.5% of the browsers support it [1]. This
percentage could be much higher, since support in 11.25% of the browsers
is unknown.

Note we still need to support the case where this field isn't supported,
and so we offer a fallback and on the server side we don't assume we're
always getting a time. We're doing a strange hack so we set the field
type to text before changing its value; otherwise old Firefox browsers
crashed.

Also note that, until now, we were storing end dates in the database as
a date with 00:00 as its time, but we were considering the poll to be
open until 23:59 that day. So, in order to keep backwards compatibility,
we're adding a task to update the dates of existing polls so we get the
same behavior we had until now.

Finally, there's one test where we're using `beginning_of_minute` to set
a pull. That's because Chrome provides an interface to enter a time in a
`%H:%M` format when the "seconds" vaule of the provided time is zero.
However, when the "seconds" value isn't zero, Chrome provides an
interface to enter a time in a `%H:%M:%S` format. Since Capybara doesn't
enter the seconds when using `fill_in` with a time, the test failed when
Capybara tried to enter a time in the `%H:%M` format when Chrome
expected a time in the `%H:%M:%S` format.

To solve this last point, an alternative would be to manually provide
the format when using `fill_in` so it includes the second.

[1] https://caniuse.com/mdn-html_elements_input_type_datetime-local