Run rake tasks on all tenants TODO: tests? Is it possible to run on just some tenants?
Sending emails for the unselected or selected investments for the last
budget are triggered manually when considered appropriate given the
status of that budget, so it doesn't make sense to send them for every
tenant at once.

Some tasks don't have to run on every tenant. The task to calculate the
TSV is only done for records which were present before we added the TSV
column, and that isn't going to happen in any tenants because we added
the TSV column before adding the tenants table. Similarly, the migration
needed for existing polls isn't necessary because there weren't any
tenants before we allowed to set the starting/ending time to polls.