Validate format of subdomains
Note we're using the `:HOST` regular expression since subdomains can
contain the same characters as domains do. This isn't 100% precise,
though, since subdomains have a maximum length of 63 characters, but is
good enough for our purposes.

We're also excluding the `.` so it isn't possible to introduce nested
subdomains or domains. The reason we're doing so is we might expand this
feature in the future to accept domains and until we do so we want to
avoid the introduction of nested subdomains which would be mistaken for
domains.