Use a different attachments folder per tenant
While this is not strictly necessary, it can help moving the data of one
tenant to a different server or removing it.

Note we're using subfolders inside the `tenants` subfolder. If we
simply used subfolder with the schema name, if the schema names were,
for instance, language codes like `es`, `en`, `it`, ... they would
conflict with the default subfolders used by Active Storage.