Use the whole module name for SentencesParser
In order for `include SentencesParser` to work with Zeitwerk, we'd have
to change the code slightly:

```
module RemoteTranslations::Microsoft
  class Client
    include SentencesParser

    # (...)
  end
end
```

This would mean changing the indentation of the whole file. While we can
do that, changing the indentation of a file makes it harder to use
commands like `git blame` or `git log` with the file, so we're doing the
change the easy way.