Move logic from key definition to views
Before this change, two important things depend on the format of each
key, where to render it in the administration panel and which kind of
interface to use for each setting. Following this strategy led us to a
very complex code, very difficult to maintain or modify. So, we do not
want to depend on the setting key structure anymore to decide how or
where to render each setting.

With this commit, we introduce methods that can easily be overriden and
allow other Consul Democracy installations to customize which settings
are rendered in which tabs as well as which settings are features.