Use a button to delete site customization images
Note that we used to have the link to delete images inside the same
<form> tag as the button to update the image. However, using a button
means we're adding a new <form> tag for the action to delete the image.
This isn't valid HTML and, in some browsers, might result in the button
sending the request to the wrong URL.

As explained in commit 5311daadf, to avoid this, we'd need to replace
`button_to` with `button_tag` in the action in order to generate a
button without a form.  Then, we could add either a `form` or a
`formaction` attribute to the button.

However, I thik it's easier to move the delete button outside the update
button <form> tag. On the minus side, since the buttons no longer share
a parent, they're harder to style, so we're sticking with a simple
approach: now the form to update the image is on the left, in a
traditional vertical layout, and the button to delete the image is an
icon and is on the right. This makes the design a bit better when using
small or medium screen, particularly on touch screens, where in the past
it was easy to touch the wrong button because they were too close to
each other.