Remove unused lib/assets folder
We use vendor/assets and app/assets; the purpose of lib/assets isn't
that clear, though. According to the Rails guides:

> lib/assets is for your own libraries' code that doesn't really fit
> into the scope of the application or those libraries which are shared
> across applications.

So it must be something for companies having several Rails applications,
which isn't our case. Furthermore, this text has been removed from the
Rails guides for version 7.1, so this folder might be a legacy folder.