Disable ActiveRecord partial inserts
As mentioned in the Rails pull request [1], the main reason for partial
inserts is no longer relevant thanks to the `ignored_columns` method
(which we haven't even needed so far).

I don't have a preference regarding this setting; we're enabling it in
order to reduce the number of settings we customize.

[1] Pull request 42769 in https://github.com/rails/rails