Move edit page view to a component
This way we can simplify setting the title and styling the link in the
header. We're also fixing the unnecessary padding introduced by the
"column" classes, which caused the header not to be aligned with the
rest of the elements surrounding it.

Note that, in the CSS, the `margin-left: auto` property needs to be
included after `@include regular-button` because that mixin overwrites
the `margin-left` property. Since we're modifying this code, we're
making it compatible with RTL text, using `$global-left` instead of
`left`.